refactor multrve
