+++ f3dasm/simulator/fenics_wrapper/problems/multimaterial_rve.py
@@ -1,3 +1,4 @@
+from f3dasm.simulator.fenics_wrapper.model.bc import PeriodicBoundary
 from f3dasm.simulator.fenics_wrapper.model.domain import Domain
 from f3dasm.simulator.fenics_wrapper.model.materials.neohookean import NeoHookean
 from f3dasm.simulator.fenics_wrapper.model.materials.svenan_kirchhoff import SVenantKirchhoff
@@ -6,40 +7,38 @@ from f3dasm.simulator.fenics_wrapper.problems.problem_base import ProblemBase
 from dolfin import *
 import numpy as np
 import copy as cp
-
+import os
 
 class MultiMaterialRVE(ProblemBase):
     """ 
         General RVE model implementation
     """
-    def __init__(self, options, domain_filename, F_macro, name=None):
+    def __init__(self, options, domain_filename, name=None):
         domain = Domain(domain_filename)
         super().__init__(options=options, name=name, domain=domain)
 
         self.bc_p = PeriodicBoundary(domain,periodicity=list(range(domain.dim)),tolerance=1e-10)   # Initialize periodic boundary conditions
-        
-        ################################
         # Mixed function space initialization with periodic boundary conditions
-        ################################
         self.W = FunctionSpace(self.domain.mesh, MixedElement([self.Ve, self.Re]), constrained_domain=self.bc_p)
 
-        self.fileResults = XDMFFile("output.xdmf")
+        self.v_, self.lamb_ = TestFunctions(self.W)     # Define test functions 
+        self.w = Function(self.W)
+
+        self.time = 0
+        
+        
+    def solve(self, F_macro, work_dir):
+        if not os.path.exists(work_dir):
+            os.mkdir(work_dir)
+        self.work_dir = work_dir
+        
+        self.fileResults = XDMFFile(os.path.join(work_dir, "output.xdmf"))
         self.fileResults.parameters["flush_output"] = True
         self.fileResults.parameters["functions_share_mesh"] = True
 
         self.convergence = True
-        self.time = 0
 
-        v_,lamb_ = TestFunctions(self.W)                # Define test functions 
-        dv, dlamb = TrialFunctions(self.W)              # Define trial functions 
-        self.w = Function(self.W)
-        
-        ################################
-        # F_macro should be defined locally because when passing in another way
-        # it gives erroneous results! So for consistancy it is defined just before
-        # passing as Constant from fenics.
-        ################################
-        F_macro = Constant(self.F_macro)                
+        F_macro = Constant(F_macro)                
 
         u,c = split(self.w)
 
@@ -53,16 +52,10 @@ class MultiMaterialRVE(ProblemBase):
         ################################
         # Variational problem definition -> Lagrangian Linear Momentum Equation
         ################################
-        self.PI = inner(self.material[0].P,nabla_grad(v_))*dx(1) + inner(self.material[1].P,nabla_grad(v_))*dx(2)  
+        self.PI = inner(self.material[0].P,nabla_grad(self.v_))*dx(1) + inner(self.material[1].P,nabla_grad(v_))*dx(2)  
         
-        self.PI += dot(lamb_,u)*dx + dot(c,v_)*dx
+        self.PI += dot(self.lamb_,u)*dx + dot(c,self.v_)*dx
 
-
-    def solve(self):
-
-        """ Method: Define solver options with your solver """
-
-        self.problem()
         prm = {"newton_solver":
                 {"absolute_tolerance":1e-7,'relative_tolerance':1e-7,'relaxation_parameter':1.0,'linear_solver' : 'mumps'}}
         try:
@@ -172,91 +165,3 @@ class MultiMaterialRVE(ProblemBase):
         F = Function(V)
         solve(a_proj==b_proj,F)
         return F
-
-class PeriodicBoundary(SubDomain):
-    """
-        GENERAL PERIODIC BOUNDARY CONDITIONS IMPLEMENTATION
-
-                              #-----------# 
-                             / |        / |
-                            /  |       /  |
-                           #----------#   |
-                           *   |      |   |
-                           *   #----------# 
-                     [1]   *  *       |  / 
-                           * * [2]    | / 
-                           **         |/  
-                           ***********# 
-                               [0]
-
-            *   : Master edges/nodes
-            -   : Slave edges/nodes 
-            [i] : directions for periodicity
-
-   """
-
-    def __init__(self,domain,periodicity=None,tolerance=DOLFIN_EPS):
-
-        """ Initialize """
-
-        super().__init__()              # Initialize the base-class (Note: the tolerance is needed for the mapping method)
-
-        ################################
-        # Get the extrema of the domain for every direction
-        ################################
-        self.mins = np.array(domain.bounds[0])          
-        self.maxs = np.array(domain.bounds[1])
-
-        self.directions = np.flatnonzero(self.maxs - self.mins)     # Mark non-zero directions
-        
-        ################################
-        # Definie periodic directions
-        ################################
-        if periodicity is None:
-            self.periodic_dir = self.directions                 
-
-        else:
-            self.periodic_dir = periodicity
-
-        self.tol = tolerance 
-        self.master = []                                # Master nodes 
-        self.map_master = []                            # Mapped master nodes
-        self.map_slave = []                             # Mapped slave nodes
-
-    def inside(self,x,on_boundary):
-        ################################
-        # Mark the master nodes as True 
-        ################################
-        x_master=False                                                   
-        if on_boundary:                                                         
-            for axis in self.periodic_dir:                                      
-                if near(x[axis],self.maxs[axis],self.tol):
-                    x_master=False                                       
-                    break
-                elif near(x[axis],self.mins[axis],self.tol):
-                    x_master=True                                        
-
-        if x_master:
-            self.master.append(cp.deepcopy(x))
-
-        return x_master
-
-    # Overwrite map method of SubDomain class to define the mapping master -> slave
-    def map(self,x,y):
-        ################################
-        # Map the master nodes to slaves
-        ################################
-        x_slave=False                                       
-        for axis in self.directions:                               
-            if axis in self.periodic_dir:                          
-                if near(x[axis],self.maxs[axis],self.tol):
-                    y[axis]=self.mins[axis]                        
-                    x_slave=True                            
-                else:                                              
-                    y[axis]=x[axis]                                
-            else:                                                  
-                y[axis]=x[axis]                                    
-
-        if x_slave:
-            self.map_master.append(cp.deepcopy(y))                            # add y to list mapped master coordinates
-            self.map_slave.append(cp.deepcopy(x))                             # add x to list of mapped slave coordinates
