+++ f3dasm/simulator/fenics_wrapper/problems/problem_base.py
@@ -10,16 +10,22 @@ class ProblemBase():
     Attributes;
         name: name of the model
         domain: mesh 
-        V: function space
-        u: trial function
-        v: test function
-        bcs: list of boundary conditions
+        Ve: function space for the problem
+        Re: function space for the pure Neumann Lagrange multiplier
+        # u: trial function
+        # v: test function
+        # bcs: list of boundary conditions
 
     """
     def __init__(self, options: dict, domain: Domain, name: Optional[str]=None) -> None:
         self.options = options
         self.name = name
         self.domain = domain
+        # Define function spaces for the problem
+        self.Ve = dolfin.VectorElement("CG", self.domain.mesh.ufl_cell(), 1)
+        # Define function space for the pure Neumann Lagrange multiplier
+        self.Re = dolfin.VectorElement("R", self.domain.mesh.ufl_cell(), 0)
+
     # bcs: Optional[List[dolfin.SubDomain]] = None
     # V: Optional[dolfin.FunctionSpace]=None
     # u: Optional[dolfin.TrialFunction] = None
